+++ lib/eve_online.rb
@@ -26,6 +26,7 @@ require 'eve_online/esi/alliance_icon'
 require 'eve_online/esi/character_assets'
 require 'eve_online/esi/character_assets_locations'
 require 'eve_online/esi/character_assets_names'
+require 'eve_online/esi/corporation_assets'
 
 require 'eve_online/esi/character_bookmarks'
 require 'eve_online/esi/character_bookmark_folders'