+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `EveOnline::ESI::CharacterAssetsLocations#etag` raise `NotImplementedError` from now
 * `EveOnline::ESI::CharacterAssetsNames#etag` raise `NotImplementedError` from now
 * Add corporation roles support
+* Add `EveOnline::ESI::CorporationAssets`
 
 ## v0.24.0